requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.25
python-dotenv==1.0.0
Werkzeug==3.0.1
requests==2.31.0
python-steam-api==2.0.4
